% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carabidae.R
\docType{data}
\name{carabidae}
\alias{carabidae}
\title{Counts of arthropods in a grid-sampled wheat field}
\format{
A matrix where each row and column is a count of carabidae species at the
first cell.
}
\source{
Kevin Wright (2020). agridat: Agricultural Datasets. R package version 1.17.
https://CRAN.R-project.org/package=agridat
}
\usage{
data("carabidae")
}
\description{
A matrix with  the counts of arthropods (carabidae) in a grid-sampled wheat
field
}
\keyword{datasets}
