% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nematodes}
\alias{nematodes}
\title{A matrix of counting data with 15 rows and 15 columns.}
\format{
Dataset of 15 × 15 regular grid on counts of cereal cyst-nematode
(Heterodera avenae collected in soil core).
}
\source{
PERRY, J. N. et al. SADIE: software to measure and model spatial pattern.
Aspects of applied biology, v. 46, p. 95-102, 1996.
}
\usage{
data("nematodes")
}
\description{
A matrix of counting data with 15 rows and 15 columns.
}
\keyword{datasets}
