% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_sinar.R
\name{var_sinar}
\alias{var_sinar}
\title{Empirical estimate for the variance of innovations.}
\usage{
var_sinar(X)
}
\arguments{
\item{X}{A integer matrix where each cell is the observed value in the
regular lattice.}
}
\value{
The estimated standard deviation in the \eqn{SINAR(1,1)}.
}
\description{
\eqn{\sigma^2_\epsilon} is the variance the innovations for the
\eqn{SINAR(1,1)} model.
}
\examples{

A simulated regular lattice with poison innovations.
n_row <- 20
n_col <- 50
a10 <- 0.2
a01 <- 0.2
a11 <-  0.5
l <- 1
X <-  sinar_pois(n_row, n_col, a10, a01, a11, l)

W <- emp_V(X)
}
