% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_sinar.R
\name{var_hat_sigma}
\alias{var_hat_sigma}
\title{Variance of standard deviation of epsilon.}
\usage{
var_hat_sigma(X)
}
\arguments{
\item{X}{A integer matrix where each cell is the observed value in the
regular lattice.}
}
\value{
The variance of standard deviation of the estimate of
\eqn{\sigma_\epsilon}.
}
\description{
\eqn{\hat{\sigma}_\epsilon} is the standard deviation of \eqn{SINAR(1,1)}
model.
}
\examples{

A simulated regular lattice with poison innovations.
n_row <- 20
n_col <- 50
a10 <- 0.2
a01 <- 0.2
a11 <-  0.5
l <- 1
X <-  sinar_pois(n_row, n_col, a10, a01, a11, l)
}
