% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_sinar.R
\name{var_sinar}
\alias{var_sinar}
\title{Empirical estimate for the variance of innovations.}
\usage{
var_sinar(X)
}
\arguments{
\item{X}{A integer matrix where each cell is the observed value in the
regular lattice.}
}
\value{
The estimated standard deviation in the \eqn{SINAR(1,1)}.
}
\description{
\eqn{\sigma^2_\epsilon} is the variance the innovations for the
\eqn{SINAR(1,1)} model.
}
\examples{

data("nematodes")
var_sinar(nematodes)
}
